refactor: streamline `CompletableFuture` waiting in `AsyncTestUtils` by adding collection support and updating test suites to use it.

+++ src/test/java/com/justsyncit/scanner/AsyncTestRunner.java
@@ -83,9 +83,9 @@ public final class AsyncTestRunner {
             }
 
             // Calculate class-level results
-            int passedCount = (int) methodResults.stream().mapToInt(r -> r.isPassed() ? 1 : 0).sum();
-            int failedCount = (int) methodResults.stream().mapToInt(r -> r.isFailed() ? 1 : 0).sum();
-            int skippedCount = (int) methodResults.stream().mapToInt(r -> r.isSkipped() ? 1 : 0).sum();
+            int passedCount = methodResults.stream().mapToInt(r -> r.isPassed() ? 1 : 0).sum();
+            int failedCount = methodResults.stream().mapToInt(r -> r.isFailed() ? 1 : 0).sum();
+            int skippedCount = methodResults.stream().mapToInt(r -> r.isSkipped() ? 1 : 0).sum();
 
             boolean classSuccess = failedCount == 0;
             Duration executionTime = Duration.between(startTime, Instant.now());
